Refactor activity processing helpers
