+++ library/processing/activity/__init__.py
@@ -1,0 +1,19 @@
+"""Activity processing utilities."""
+
+from .annotations import (
+    apply_activity_annotations,
+    build_activity_properties,
+    extract_effect_features,
+    infer_action_type,
+    normalise_mapping,
+)
+from .bounds import compute_activity_bounds
+
+__all__ = [
+    "apply_activity_annotations",
+    "build_activity_properties",
+    "compute_activity_bounds",
+    "extract_effect_features",
+    "infer_action_type",
+    "normalise_mapping",
+]
